master separation branch:   - separated this branch to work on frontend + backend + mongoDB   - several fixes:     - removed unnecessary event callback from onClick events in CartPAge. replaced them only with the reference of prop clickQuantity     - separated App from index.js in a different component called App     - updated README.md to reflect both branches   - added StrictMode to wrap over App   - fixed prototype pollutions severity warnings on the frontend side (except for yargs-parser package)
